finsh calculate community metrics refactoring
